2026-04-11-unconf-sf/pyproject.toml
[project]
name = "unconf-sf-transcriber"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.9.9",
    "baml-py==0.220.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
